refactor: consolidate replacement logic
