+++ src/agent/modelHandlers/modelHandlerGoogleGenAI.ts
@@ -34,8 +34,8 @@ import { AgentLogger } from '@logger/AgentLogger';
 // Local imports - utilities
 import { WorkspaceFS, AbsoluteFS } from '@utils/files';
 import { formatProviderError } from '@utils/sdkErrorUtils';
-import { cleanFileContent } from '@replacement/replacementUtils';
-import replacementManager from '@replacement/replacementManager';
+import { cleanFileContent } from '@replacement/engine';
+import replacementEngine from '@replacement/engine';
 import { extractScratchpad } from '@utils/text/xmlUtils';
 import type { ProviderStopReason } from '../../types/StopReasonTypes';
 import { getConfig } from '@utils/config';
@@ -607,7 +607,7 @@ export class ModelHandlerGoogleGenAI extends ModelHandler {
       );
     }
 
-    responseText = replacementManager.applyAll(responseText);
+    responseText = replacementEngine.applyAll(responseText);
 
     const usage = responseObject.usageMetadata;
     const stopReason: FinishReason =
