+++ src/replacementEngine/advanced.ts
