+++ src/replacementEngine/engine.ts
@@ -11,12 +11,12 @@ import { getConfig } from '@utils/config';
 const CHANNEL = 'ReplacementUtils';
 logger.initialize(CHANNEL);
 
-import { ReplacementCategory } from './replacementTypes';
+import { ReplacementCategory } from './types';
 import {
   applyLatexQuotesFormatting,
   replaceMathUnicode,
   fixLatexQuoteIssues,
-} from './replacementAdvanced';
+} from './advanced';
 
 import {
   EQUATION_REPLACEMENTS,
@@ -31,18 +31,15 @@ import {
   GPTNESS_REPLACEMENTS,
   PERSONAL_STYLE_REPLACEMENTS,
   LATEXDIFF_REPLACEMENTS,
-} from './replacementRules';
-import {
-  MAX_STYLE_REPLACEMENTS,
-  MAX_REGEX_REPLACEMENTS,
-} from './replacementMax';
+} from './rules';
+import { MAX_STYLE_REPLACEMENTS, MAX_REGEX_REPLACEMENTS } from './maxRules';
 
 import {
   PARENTHESES_REPLACEMENTS,
   LATEXDIFF_MARKUP_REPLACEMENTS,
   INLINE_MATH_REPLACEMENTS,
   EQUATION_STYLE_REPLACEMENTS,
-} from './replacementRulesRegex';
+} from './rulesRegex';
 
 // ===== LaTeX Content Formatting =====
 
@@ -273,3 +270,35 @@ export function cleanFileContent(content: string): string {
   cleaned = applyReplacements(cleaned, getAllReplacementsRegex()).trim();
   return cleaned;
 }
+
+/**
+ * Provides high-level APIs for applying text replacement rules.
+ */
+export const replacementEngine = {
+  /**
+   * Apply all non-regex replacement rules.
+   */
+  applyNonRegex(text: string): string {
+    return applyReplacements(text, getAllReplacements()).trim();
+  },
+
+  /**
+   * Apply all regex-based replacement rules.
+   */
+  applyRegex(text: string): string {
+    return applyReplacements(text, getAllReplacementsRegex()).trim();
+  },
+
+  /**
+   * Apply the full set of replacement rules. Non-regex replacements are
+   * executed before and after regex replacements to resolve any nested
+   * transformations.
+   */
+  applyAll(text: string): string {
+    let processed = this.applyNonRegex(text);
+    processed = this.applyRegex(processed);
+    return this.applyNonRegex(processed);
+  },
+};
+
+export default replacementEngine;
