+++ src/replacementEngine/helpers.ts
@@ -3,144 +3,13 @@
  */
 import { capitalize } from '@frontend/ui/messageUtils';
 
-// ===== Shared Constants =====
-
-// Greek Letters
-// prettier-ignore
-export const GREEK_LETTERS = 'alpha beta gamma delta epsilon zeta eta theta iota kappa lambda mu nu xi omicron pi rho sigma tau upsilon phi chi psi omega'.split(' ');
-
-// prettier-ignore
-export const GREEK_UPPERCASE = 'Alpha Beta Gamma Delta Epsilon Zeta Eta Theta Iota Kappa Lambda Mu Nu Xi Omicron Pi Rho Sigma Tau Upsilon Phi Chi Psi Omega'.split(' ');
-
-// Common LaTeX environments
-// prettier-ignore
-export const COMMON_ENVIRONMENTS = [
-  // 'document',
-  'figure',
-  'figure*',
-  'tikzpicture',
-  'align',
-  'equation',
-  'itemize',
-  'enumerate',
-  'tabular',
-  'abstract',
-  'theorem',
-  'proof',
-  'definition',
-  'corollary',
-  'lemma',
-  'proposition',
-  'remark',
-  'example',
-  'exercise',
-  'problem',
-  'solution',
-  'acknowledgment',
-
-  //
-  'minipage',
-  'verbatim',
-  'lstlisting',
-  'minted',
-  'algorithm',
-  'tcolorbox',
-
-  // beamer:
-  'frame',
-  // 'block',
-  'alertblock',
-  'exampleblock',
-];
-
-// Math Environments
-// prettier-ignore
-export const MATH_ENVIRONMENTS = [
-  'align',
-  'align*',
-  'equation',
-  'equation*',
-  'aligned',
-  'aligned*',
-  'gather',
-  'gather*',
-  'multline',
-  'split',
-  'cases',
-  'alignat',
-  'alignat*',
-  'flalign',
-  'flalign*',
-  'pmatrix',
-  'bmatrix',
-  'vmatrix',
-  'Vmatrix',
-  // 'array', // avoid super general names
-  // 'array*',
-];
-
-// Section types
-export const SECTION_TYPES = [
-  'chapter',
-  'section',
-  'subsection',
-  'subsubsection',
-  'paragraph',
-];
-
-// Math operators
-// prettier-ignore
-export const MATH_OPERATORS = [
-  'cos',
-  'sin',
-  'tan',
-  'arctan',
-  'arccos',
-  'arcsin',
-  'log',
-  'ln',
-  'exp',
-  'sqrt',
-  'max',
-  'min',
-  'sup',
-  'inf',
-  'lim',
-  'csc',
-  'ker',
-  'limsup',
-  'deg',
-  'gcd',
-  'lg',
-  'Pr',
-  'cot',
-  'det',
-  'hom',
-  'sec',
-  'arg',
-  'coth',
-  'dim',
-  'liminf',
-];
-
-// prettier-ignore
-export const DOTS = 'cdot dot ldots cdots vdots ddots iddots'.split(' ');
-export const INTEGRALS = 'int iint iiint oint ooint ooooint'.split(' ');
-
-// Reference types
-// prettier-ignore
-export const REFERENCE_TYPES = [
-  'figure',
-  'table',
-  'equation',
-  'section',
-  'fig.',
-  'eq.',
-  'eqn.',
-  'Eq.',
-  'Eqs.',
-  'Fig.',
-];
+// Re-export commonly used constants
+export {
+  GREEK_LETTERS,
+  SECTION_TYPES,
+  MATH_OPERATORS,
+  LATEX_ENVIRONMENTS,
+} from './constants';
 
 /**
  * Generates patterns for fixing extra backslashes in LaTeX commands
