+++ tsconfig.json
@@ -22,7 +22,7 @@
       "@housekeeping/*": ["src/housekeeping/*"],
       "@progressView/*": ["src/progressView/*"],
       "@historyView/*": ["src/historyView/*"],
-      "@replacement/*": ["src/replacement/*"],
+      "@replacement/*": ["src/replacementEngine/*"],
       "@tools/*": ["src/tools/*"],
       "@types/*": ["src/types/*"]
     },
